Add files via upload
This image test.jpg most likely belongs to Cat with a 73.10 percent confidence.
This image test2.jpg most likely belongs to Dog with a 73.11 percent confidence.
This image notadog.jpg most likely belongs to Dog with a 73.11 percent confidence.
This image notadogagain.jpg most likely belongs to Cat with a 70.95 percent confidence.
This image bestfriends?.jpg most likely belongs to Dog with a 73.09 percent confidence.
This image dog?.JPG most likely belongs to Cat with a 68.65 percent confidence.
WARNING:tensorflow:No training configuration found in the save file, so the model was *not* compiled. Compile it manually.
/home/domenico/PycharmProjects/hw4/classify.py:65: DeprecationWarning: ANTIALIAS is deprecated and will be removed in Pillow 10 (2023-07-01). Use Resampling.LANCZOS instead.
  img = img.resize((new_width, new_height), Image.ANTIALIAS)
We got:  0.994
[[9.999912e-01 8.787405e-06]]
[0]
tf.Tensor([0.7310551 0.2689449], shape=(2,), dtype=float32)
This image most likely belongs to Cat with a 73.11 percent confidence.
WARNING:tensorflow:No training configuration found in the save file, so the model was *not* compiled. Compile it manually.
/home/domenico/PycharmProjects/hw4/classify.py:115: DeprecationWarning: ANTIALIAS is deprecated and will be removed in Pillow 10 (2023-07-01). Use Resampling.LANCZOS instead.
  img = img.resize((new_width, new_height), Image.ANTIALIAS)